Time: 122 ms (76.78%), Space: 63 MB (60.66%) - LeetHub

+++ 2616-maximal-score-after-applying-k-operations/2616-maximal-score-after-applying-k-operations.java
@@ -6,7 +6,7 @@ class Solution {
         while(k!=0 && !pq.isEmpty()) {
             int current = pq.poll();
             result += current;
-            pq.offer((int)Math.ceil(current/3.0));
+            pq.offer((int)Math.ceil(current/(double)3));
             k--;
         }
         return result;
